libnodeupdown: remove openib backend module
Problem: The openib backend module has not been updated since 2009
or so.  It currently segfaults on most systems as the API has changed.
Some mild attempts to get it working have shown little progress
as it's unclear how the API / protocol have changed.

Lets remove the module as its not working anyways.